Make cloudflare-zlib dependency optional
Although the Cloudflare fork of zlib may introduce performance
improvements, there are issues building it with MinGW toolchains on
Windows, as evidenced by this StackOverflow question: https://stackoverflow.com/questions/61170244/gcc-exe-fail-when-building-cloudflare-zlib-sys-rs

However, it is not always possible to switch to the MSVC toolchain on
Windows, because that may break other native C libraries which have
trouble with MSVC instead, and just compiling to a 32-bit target can get
pretty messy too.

I've opened an issue to get this error fixed upstream, in the
cloudflare-zlib-sys crate, but it was not fixed there yet.

Therefore, make the Cloudflare zlib dependency optional, so users will
have to explicitly opt in for it. As a side bonus, these changes also
remove the restriction of Cloudflare zlib only being available on x64
and Aarch64 platforms.